+++ test/Sdk.Generator.Tests/FunctionMetadataProvider/DiagnosticResultTests.cs
@@ -259,6 +259,112 @@ namespace Microsoft.Azure.Functions.Sdk.Generator.FunctionMetadataProvider.Tests
                     expectedDiagnosticResults,
                     languageVersion: languageVersion);
             }
+
+            [Theory]
+            [InlineData(LanguageVersion.CSharp7_3)]
+            [InlineData(LanguageVersion.CSharp8)]
+            [InlineData(LanguageVersion.CSharp9)]
+            [InlineData(LanguageVersion.CSharp10)]
+            [InlineData(LanguageVersion.CSharp11)]
+            [InlineData(LanguageVersion.Latest)]
+            public async Task DuplicateFunctionNamesFails(LanguageVersion languageVersion)
+            {
+                var inputCode = @"using System;
+                    using System.Threading.Tasks;
+                    using Microsoft.Azure.Functions.Worker;
+                    using Microsoft.Azure.Functions.Worker.Http;
+
+                    namespace FunctionApp
+                    {
+                        public class FunctionClassA
+                        {
+                            [Function(""MyFunc"")]
+                            public string RunA([HttpTrigger(""get"")] string req)
+                            {
+                                throw new NotImplementedException();
+                            }
+                        }
+
+                        public class FunctionClassB
+                        {
+                            [Function(""MyFunc"")]
+                            public string RunB([HttpTrigger(""get"")] string req)
+                            {
+                                throw new NotImplementedException();
+                            }
+                        }
+                    }";
+
+                string? expectedGeneratedFileName = null;
+                string? expectedOutput = null;
+
+                var expectedDiagnosticResults = new List<DiagnosticResult>
+                {
+                    new DiagnosticResult(DiagnosticDescriptors.DuplicateFunctionName)
+                        .WithArguments("MyFunc", "FunctionApp.FunctionClassA.RunA, FunctionApp.FunctionClassB.RunB")
+                };
+
+                await TestHelpers.RunTestAsync<FunctionMetadataProviderGenerator>(
+                    referencedExtensionAssemblies,
+                    inputCode,
+                    expectedGeneratedFileName,
+                    expectedOutput,
+                    expectedDiagnosticResults,
+                    languageVersion: languageVersion);
+            }
+
+            [Theory]
+            [InlineData(LanguageVersion.CSharp7_3)]
+            [InlineData(LanguageVersion.CSharp8)]
+            [InlineData(LanguageVersion.CSharp9)]
+            [InlineData(LanguageVersion.CSharp10)]
+            [InlineData(LanguageVersion.CSharp11)]
+            [InlineData(LanguageVersion.Latest)]
+            public async Task DuplicateFunctionNamesCaseInsensitiveFails(LanguageVersion languageVersion)
+            {
+                var inputCode = @"using System;
+                    using System.Threading.Tasks;
+                    using Microsoft.Azure.Functions.Worker;
+                    using Microsoft.Azure.Functions.Worker.Http;
+
+                    namespace FunctionApp
+                    {
+                        public class FunctionClassA
+                        {
+                            [Function(""MyFunc"")]
+                            public string RunA([HttpTrigger(""get"")] string req)
+                            {
+                                throw new NotImplementedException();
+                            }
+                        }
+
+                        public class FunctionClassB
+                        {
+                            [Function(""myfunc"")]
+                            public string RunB([HttpTrigger(""get"")] string req)
+                            {
+                                throw new NotImplementedException();
+                            }
+                        }
+                    }";
+
+                string? expectedGeneratedFileName = null;
+                string? expectedOutput = null;
+
+                var expectedDiagnosticResults = new List<DiagnosticResult>
+                {
+                    new DiagnosticResult(DiagnosticDescriptors.DuplicateFunctionName)
+                        .WithArguments("MyFunc", "FunctionApp.FunctionClassA.RunA, FunctionApp.FunctionClassB.RunB")
+                };
+
+                await TestHelpers.RunTestAsync<FunctionMetadataProviderGenerator>(
+                    referencedExtensionAssemblies,
+                    inputCode,
+                    expectedGeneratedFileName,
+                    expectedOutput,
+                    expectedDiagnosticResults,
+                    languageVersion: languageVersion);
+            }
         }
     }
 }